Sections Added
Different sections of the web page are now divided up into div tags, each one (currently) featuring a header, a figure/figcaption tag, and a paragraph.

A problem has arisen: Depending on the length of the paragraph, the figure sometimes "sticks out" of the div, overlapping with other content. Let's see if we can get it to not do that, please?