Nav Bar
Updated the contact page slightly (Not super important right now).

More importantly, navbar.html has the template for the navbars that should, ideally, be used on every page, if it works out that way. It's almost entirely bootstrap stuff, but also uses the Font Awesome stylesheet, so let's be sure to add that to the other pages as well.